+++ arc_application/views/nanny_views/nanny_arc_summary.py
@@ -16,7 +16,7 @@ from .nanny_childcare_training import NannyChildcareTrainingSummary
 from .nanny_dbs_check import NannyDbsCheckSummary
 from .nanny_insurance_cover import NannyInsuranceCoverSummary
 
-from arc_application.views.nanny_views.nanny_view_helpers import nanny_all_completed
+from arc_application.services.nanny_view_helpers import nanny_all_completed
 from arc_application.views.nanny_views.nanny_email_helpers import send_accepted_email, send_returned_email
