+++ arc_application/views/nanny_views/nanny_personal_details.py
@@ -2,7 +2,7 @@ from ...forms.nanny_forms.nanny_form_builder import HomeAddressForm, PersonalDet
 from ...services.db_gateways import NannyGatewayActions
 from .nanny_form_view import NannyARCFormView
 
-from .nanny_view_helpers import parse_date_of_birth
+from arc_application.services.nanny_view_helpers import parse_date_of_birth
 
 
 class NannyPersonalDetailsSummary(NannyARCFormView):
